Extract InitializerReturnValueBuilder to its own file
